+++ src/core/src/platform.rs
@@ -3,6 +3,19 @@ use platforms::{
     TARGET_ARCH, TARGET_OS,
 };
 
+/// Represents an endianness of a CPU.
+#[repr(C)]
+#[allow(non_camel_case_types)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum NSTDEndian {
+    /// An unknown-endian.
+    NSTD_ENDIAN_UNKNOWN,
+    /// Little-endian.
+    NSTD_ENDIAN_LITTLE,
+    /// Big-endian.
+    NSTD_ENDIAN_BIG,
+}
+
 /// Represents a CPU architecture.
 #[repr(C)]
 #[non_exhaustive]
@@ -173,6 +186,19 @@ pub struct NSTDPlatform {
     pub arch: NSTDCPUArch,
     /// The operating system.
     pub os: NSTDOperatingSystem,
+}
+
+/// Returns the target CPU's endianness.
+/// Returns: `NSTDEndian endian` - The target CPU endianness.
+#[inline]
+#[cfg_attr(feature = "clib", no_mangle)]
+pub unsafe extern "C" fn nstd_core_platform_endian() -> NSTDEndian {
+    #[cfg(target_endian = "little")]
+    return NSTDEndian::NSTD_ENDIAN_LITTLE;
+    #[cfg(target_endian = "big")]
+    return NSTDEndian::NSTD_ENDIAN_BIG;
+    #[cfg(not(any(target_endian = "little", target_endian = "big")))]
+    return NSTDEndian::NSTD_ENDIAN_UNKNOWN;
 }
 
 /// Returns an `NSTDCPUArch` value representing the target CPU architecture.